edited building images and tenante page
building images now are smaller and easier to see, will need to edit the images to be a certain size so there isnt weird white space. The tenant page has been edited to just pull from the database and be able to be updated with however many are in the database to call on.