created toast and carasoul components
the toast component will notify the user based on whatever we set it to, as of right now it needs to be adjusted to be based on when a user submits a message, and the carasoul is being designed to allow for a nicer view of our building listing. current issue with it is needs to be connected to the designated page and then have the necessary images added to it. The library specifies how to use it with props so it can be scaled to show our users own added images if we set it up that way.